+++ resources/release-notes.md
@@ -11,3 +11,7 @@
 - Move `cropResults` definition from `mag.app.control.Control` to `mag.app.internal.cropResults`
 - Redesign `mag.app.control.StartEndDate/addStartEndDateButtons` signature to be more flexible
 - Redesign `mag.app.control.Filter/addFilterButtons` signature to be more flexible
+
+## Tests
+
+- Add tests for `mag.app.internal.combineDateAndTime`